requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
pytest==5.3.2
